Implements ownership verification for to-do delete and update

+++ backend/src/services/todo_service.py
@@ -1,4 +1,4 @@
-from fastapi import Depends
+from fastapi import Depends, HTTPException, status
 from src.repositories import TodoRepository
 from src.schemas.request import CreateTodoRequestSchema, UpdateTodoRequestSchema
 from src.models import Todo
@@ -15,14 +15,30 @@ class TodoService:
 
         return inserted_todo
 
-    def update_todo(self, todo: UpdateTodoRequestSchema) -> Todo:
-        todo_to_update = Todo(**todo.model_dump())
+    def update_todo(self, user_id: int, todo: UpdateTodoRequestSchema) -> Todo:
+        todo_to_update = self.repository.get(todo.id)
+
+        if not todo_to_update:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='To-do not found')
+
+        if todo_to_update.user_id != user_id:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='To-do doesn\'t belong to user')
+
+        todo_to_update.description = todo.description
 
         updated_todo = self.repository.update(todo_to_update)
 
         return updated_todo
     
-    def delete_todo(self, todo_id: int) -> bool:
+    def delete_todo(self, user_id: int, todo_id: int) -> bool:
+        todo_to_delete = self.repository.get(todo_id)
+
+        if not todo_to_delete:
+            return False
+
+        if todo_to_delete.user_id != user_id:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='To-do doesn\'t belong to user')
+
         return self.repository.delete(todo_id)
 
     def get_todos_by_user_id(self, user_id: int) -> list[Todo]:
